Update main.cpp
Fixed the issue with the creating objects with an index. The new code should be able to do any of our functions with no problem(Which I tested for a few different functions and seemed to be the case) as long as you have the proper index for the student you are trying to do.